Changed bell sensor function...
Made it so that the bell sensor can be used to close the pipe with any fingering pattern, i.e. if all holes and the bell sensor are covered (and you're not using the bell sensor for controlling the register) then no note will sound. This was requested by Gaita players who close the pipe similar to uilleann players. I don't think it can hurt to have it on all fingering patterns.